fix(Slice): cast begin_/end_ to int32 before clip_by_value
ONNX allows Slice indices to be int64.  When begin_ or end_ are int64
tensors, tf.subtract(1, tf.clip_by_value(t=begin_, ...)) fails with:

  TypeError: Input 'y' of 'Sub' Op has type int64 that does not match
  type int32

because the literal 1 is inferred as int32 while begin_/end_ are int64.

Fix: cast begin_ and end_ to int32 before passing them to clip_by_value,
consistent with the surrounding int32 arithmetic.

+++ onnx2tf/ops/Slice.py
@@ -491,7 +491,7 @@ def make_node(
             cliped_values = tf.cast(
                 tf.subtract(
                     1,
-                    tf.clip_by_value(t=begin_, clip_value_min=0, clip_value_max=1),
+                    tf.clip_by_value(t=tf.cast(begin_, tf.int32), clip_value_min=0, clip_value_max=1),
                 ),
                 dtype=tf.int32,
             )
@@ -529,7 +529,7 @@ def make_node(
             cliped_values = tf.cast(
                 tf.subtract(
                     1,
-                    tf.clip_by_value(t=end_, clip_value_min=0, clip_value_max=1),
+                    tf.clip_by_value(t=tf.cast(end_, tf.int32), clip_value_min=0, clip_value_max=1),
                 ),
                 dtype=tf.int32,
             )
